# src/main/webapp/docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:8089/api/v1
  - url: https://todolist-api-348718.ew.r.appspot.com/api/v1
info:
  description: Crea tu propia lista de tareas para cualquier proyecto.
  version: 1.0.0
  title: TODOlist-API
  contact:
    email: laurolmer@us.es
...
paths:
...
  '/tasks/{taskId}':
    get:
      tags:
        - Tasks
      summary: Get task
      description: Get a task
      operationId: getTask
      parameters:
        - name: taskId
          in: path
          description: Id of the task
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete task
      operationId: deleteTask
      parameters:
        - name: taskId
          in: path
          description: Id of the task
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Task deleted
        '404':
          description: Task not found
...
  '/users/{userId}':
    get:
      tags:
        - User
      summary: Get an user
      description: Gets an user
      operationId: getUser
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
        - name: fieldsUser
          in: query
          description: Attributes of a User
          required: false
          schema:
            type: string
        - name: fieldsTask
          in: query
          description: Attributes of a task
          required: false
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        '404':
          description: User not found
    delete:
      tags:
        - User
      summary: Delete user
      description: Delete user
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found
  '/users/{userId}/{taskId}':
    post:
      tags:
        - User
      summary: Add a task to a user
      description: Add a task to a user
      operationId: addTaskToUser
      requestBody:
        content:
          text/plain:
            schema:
              $ref: '#/components/schemas/User'
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          description: Id of the task
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Task added to the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '404':
          description: User or task not found
    delete:
      tags:
        - User
      summary: Delete task from user
      description: Delete task from user
      operationId: deleteTaskToUser
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          description: Id of the task
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Task deleted from user
        '404':
          description: User or task not found
  /github/{account}:
    get:
      tags:
        - GitHub User
      summary: Get GitHub user
      description: Gets a GitHub user
      operationId: getGitHubUser
      parameters:
        - name: account
          in: path
          description: Name of the GitHub account
          required: true
          schema:
            type: string
      responses:
        '200':
          description: GitHub user found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: GitHub user not found
    post:
      tags:
        - GitHub User
      summary: Add GitHub user
      description: Adds a GitHub user
      operationId: addGitHubUser
      requestBody:
        content:
          text/plain:
            schema:
              $ref: '#/components/schemas/User'
      parameters:
        - name: account
          in: path
          description: Name of the GitHub account
          required: true
          schema:
            type: string
      responses:
        '201':
          description: GitHub user added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '404':
          description: GitHub user not found
  '/github/{account}/{repo}':
    get:
      tags:
        - GitHub User
      summary: Get task from GitHub user
      description: Gets a task from a GitHub user
      operationId: getTaskGitHubUser
      parameters:
        - name: account
          in: path
          description: Name of a GitHub account
          required: true
          schema:
            type: string
        - name: repo
          in: path
          description: Name of a GitHub repository
          required: true
          schema:
            type: string
        - name: status
          in: query
          description: Status of a task
          required: false
          schema:
            type: string
        - name: finishedDate
          in: query
          description: When the task finished (after start date)
          required: false
          schema:
            type: string
        - name: priority
          in: query
          description: Priority of a task (between 0 and 5)
          required: false
          schema:
            type: string
        - name: difficulty
          in: query
          description: Difficulty of a task (it could be EASY, MEDIUM, HARD, HARDCORE and I WANT TO DIE)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: GitHub user found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: GitHub user not found
    post:
      tags:
        - GitHub User
      summary: Add task to GitHub user
      description: Adds a task to a GitHub user
      operationId: addTaskGitHubUser
      requestBody:
        content:
          text/plain:
            schema:
              $ref: '#/components/schemas/User'
      parameters:
        - name: account
          in: path
          description: Name of a GitHub account
          required: true
          schema:
            type: string
        - name: repo
          in: path
          description: Name of a GitHub repository
          required: true
          schema:
            type: string
        - name: status
          in: query
          description: Status of a task
          required: false
          schema:
            type: string
        - name: finishedDate
          in: query
          description: When the task finished (after start date)
          required: false
          schema:
            type: string
        - name: priority
          in: query
          description: Priority of a task (between 0 and 5)
          required: false
          schema:
            type: string
        - name: difficulty
          in: query
          description: Difficulty of a task (it could be EASY, MEDIUM, HARD, HARDCORE and I WANT TO DIE)
          required: false
          schema:
            type: string
      responses:
        '201':
          description: Task added to the GitHub user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '404':
          description: GitHub user or task not found
components:
  schemas:
    Task:
      type: object
      required:
        - title
      properties:
        idTask:
          description: Id of the task (is automatically generated)
          type: string
        title:
          description: Title of the task (it must be less than 50 characters)
          type: string
        description:
          description: Description of the task (it must be less than 200 characters)
          type: string
        status:
          description: Status of the task (it could be DRAFT, IN_PROGRESS, IN_REVISION, DONE, CANCELLED)
          type: string
        finishedDate:
          description: When the task finished (the following pattern should be used YYYY-MM-DD)
          type: string
        startedDate:
          description: When the task started (the following pattern should be used YYYY-MM-DD)
          type: string
        annotation:
          description: Annotation of the task (it must be less than 50 characters)
          type: string
        priority:
          description: Priority of the task (between 0 and 5)
          type: integer
        difficulty:
          description: Difficulty of the task (it could be EASY, MEDIUM, HARD, HARDCORE and I WANT TO DIE)
          type: string
        duration:
          description: Duration of the task (it's calculated using the start and finish date)
          type: integer
        taskCompleted:
          description: If the task is completed (no negatives numbers)
          type: integer
      example:
        idTask: t1
        title: Comprar el pan
        description: Ve a comprar el pan para almorzar hoy
        status: IN_PROGRESS
        finishedDate: '2022-12-12'
        startedDate: '2022-12-27'
        annotation: La panadería abre a las 9
        priority: '10'
        difficulty: EASY
        duration: '15'
    User:
      type: object
      required:
        - name
      properties:
        idUser:
          description: Id of the user (is automatically generated)
          type: string
        name:
          description: Name of the user (it must be less than 50 characters)
          type: string
        surname:
          description: Surname of the user (it must be less than 50 characters)
          type: string
        email:
          description: Email of the user (has a validation system, put a mail that makes sense)
          type: string
        avatar:
          description: Avatar of the user (has a validation system, put an URL that makes sense)
          type: string
        bio:
          description: Bio of the user (it must be less than 500 characters)
          type: string
        location:
          description: Location of the user (it must be less than 50 characters)
          type: string
        tasks:
          description: Tasks of the user
          type: array
          items:
            type: object
            required:
              - title
            properties:
              idTask:
                description: Id of the task (is automatically generated)
                type: string
              title:
                description: Title of the task (it must be less than 50 characters)
                type: string
              description:
                description: Description of the task (it must be less than 200 characters)
                type: string
              status:
                description: Status of the task (it could be DRAFT, IN_PROGRESS, IN_REVISION, DONE, CANCELLED)
                type: string
              finishedDate:
                description: When the task finished (the following pattern should be used YYYY-MM-DD)
                type: string
              startedDate:
                description: When the task started (the following pattern should be used YYYY-MM-DD)
                type: string
              annotation:
                description: Annotation of the task (it must be less than 50 characters)
                type: string
              priority:
                description: Priority of the task (between 0 and 5)
                type: integer
              difficulty:
                description: Difficulty of the task (it could be EASY, MEDIUM, HARD, HARDCORE and I WANT TO DIE)
                type: string
              duration:
                description: Duration of the task (it's calculated using the start and finish date)
                type: integer
              taskCompleted:
                description: If the task is completed (no negatives numbers)
                type: integer
      example:
        idUser: u1
        name: Kujo
        surname: Jotaro
        email: oraora@gmail.com
        avatar: https://codigoesports.com/wp-content/uploads/2019/07/quien-es-belle-delphine.jpg
        bio: Yare yare daze
        location: Morioh
        tasks:
          - idTask: t1
            title: Derrotar a Dio
            description: Cruza todo el mundo hasta llegar a Egipto
            status: DONE
            finishedDate: '2023-10-08'
            startedDate: '2022-12-12'
            annotation: Te encontrarás a un montón de randoms
            priority: '10'
            difficulty: HARDCORE
            duration: '300'
          - idTask: t2
            title: Parar a Kira
            description: Paren al psicópata explosivo
            status: DONE
            finishedDate: '2023-05-11'
            startedDate: '2022-12-12'
            annotation: Cuidado con Bites the Dust
            priority: '10'
            difficulty: HARDCORE
            duration: '150'
        taskCompleted: '2'